- Move function to build link property - Use Material icon & not fontawesome
