+++ package.json
@@ -27,6 +27,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@mdx-js/react": "^2.3.0",
+    "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.13",
     "@rebass/forms": "^4.0.6",
     "@rebass/preset": "^4.0.5",
